# Commerce-Hub/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TradeRoute B2B marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders relevant to the caller
      parameters:
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order (shopkeeper)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderInput"
      responses:
        "200":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}/status:
    post:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Manufacturer accepts/rejects/confirms an order
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}/assign-transporter:
    post:
      operationId: assignTransporter
      tags: [orders]
      summary: Manufacturer or admin assigns a transporter to an accepted order
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignTransporterInput"
      responses:
        "200":
          description: Shipment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"
...
components:
  schemas:
...
    OrderStatus:
      type: string
      enum: [pending, accepted, rejected, in_transit, delivered, cancelled]
    OrderItem:
      type: object
      required: [productId, productName, quantity, pricePerUnit, lineTotal]
      properties:
        productId:
          type: string
        productName:
          type: string
        quantity:
          type: integer
        pricePerUnit:
          type: number
        lineTotal:
          type: number
    Order:
      type: object
      required:
        - id
        - orderNumber
        - status
        - shopkeeperId
        - shopkeeperName
        - shopkeeperCompany
        - manufacturerId
        - manufacturerName
        - items
        - totalAmount
        - shippingAddress
        - createdAt
      properties:
        id:
          type: string
        orderNumber:
          type: string
        status:
          $ref: "#/components/schemas/OrderStatus"
        shopkeeperId:
          type: string
        shopkeeperName:
          type: string
        shopkeeperCompany:
          type: string
        manufacturerId:
          type: string
        manufacturerName:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        totalAmount:
          type: number
        shippingAddress:
          type: string
        notes:
          type: string
          nullable: true
        shopkeeperPhone:
          type: string
          nullable: true
        manufacturerPhone:
          type: string
          nullable: true
        shipment:
          $ref: "#/components/schemas/Shipment"
          nullable: true
        createdAt:
          type: string
          format: date-time
    CreateOrderItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: string
        quantity:
          type: integer
    CreateOrderInput:
      type: object
      required: [items, shippingAddress]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateOrderItemInput"
        shippingAddress:
          type: string
        notes:
          type: string
          nullable: true
    UpdateOrderStatusInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [accepted, rejected, cancelled]
        reason:
          type: string
          nullable: true
    AssignTransporterInput:
      type: object
      required: [transporterId]
      properties:
        transporterId:
          type: string
        pickupDate:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
...
    ShipmentStatus:
      type: string
      enum: [assigned, picked_up, in_transit, delivered]
    Shipment:
      type: object
      required:
        - id
        - trackingNumber
        - status
        - orderId
        - orderNumber
        - transporterId
        - transporterName
        - originAddress
        - destinationAddress
        - createdAt
      properties:
        id:
          type: string
        trackingNumber:
          type: string
        status:
          $ref: "#/components/schemas/ShipmentStatus"
        orderId:
          type: string
        orderNumber:
          type: string
        transporterId:
          type: string
        transporterName:
          type: string
        transporterCompany:
          type: string
          nullable: true
        shopkeeperName:
          type: string
          nullable: true
        shopkeeperCompany:
          type: string
          nullable: true
        shopkeeperPhone:
          type: string
          nullable: true
        manufacturerName:
          type: string
          nullable: true
        manufacturerCompany:
          type: string
          nullable: true
        manufacturerPhone:
          type: string
          nullable: true
        originAddress:
          type: string
        destinationAddress:
          type: string
        pickupDate:
          type: string
          format: date-time
          nullable: true
        deliveredAt:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        originLat:
          type: number
          nullable: true
        originLng:
          type: number
          nullable: true
        destLat:
          type: number
          nullable: true
        destLng:
          type: number
          nullable: true
        currentLat:
          type: number
          nullable: true
        currentLng:
          type: number
          nullable: true
        lastLocationAt:
          type: string
          format: date-time
          nullable: true
        pickupPin:
          type: string
          nullable: true
        deliveryPin:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
...