# Commerce-Hub/lib/api-zod/src/generated/types/orderStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TradeRoute B2B marketplace API
 * OpenAPI spec version: 0.1.0
 */

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  accepted: "accepted",
  rejected: "rejected",
  in_transit: "in_transit",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;